# Remove chmod calls for running scripts