--- a/perfmetrics/scripts/continuous_test/ml_tests/build.sh
+++ b/perfmetrics/scripts/continuous_test/ml_tests/build.sh
@@ -14,7 +14,6 @@ source ~/.bashrc
 cd "${KOKORO_ARTIFACTS_DIR}/github/gcsfuse/perfmetrics/scripts"
 
 echo "Setting up a machine"
-chmod +x ml_tests/setup.sh
 source ml_tests/setup.sh
 
 echo "Running ML model automation script"