--- a/perfmetrics/scripts/continuous_test/ml_tests/tf/resnet/build.sh
+++ b/perfmetrics/scripts/continuous_test/ml_tests/tf/resnet/build.sh
@@ -10,6 +10,5 @@ TEST_SCRIPT_PATH="github/gcsfuse/perfmetrics/scripts/ml_tests/tf/resnet/setup_ho
 
 cd "${KOKORO_ARTIFACTS_DIR}/github/gcsfuse/perfmetrics/scripts/continuous_test/ml_tests/"
 
-chmod +x run_and_manage_test.sh
 source run_and_manage_test.sh $VM_NAME $ZONE_NAME $ARTIFACTS_BUCKET_PATH $TEST_SCRIPT_PATH