--- a/perfmetrics/scripts/ml_tests/pytorch/dino/setup_host_and_run_model.sh
+++ b/perfmetrics/scripts/ml_tests/pytorch/dino/setup_host_and_run_model.sh
@@ -6,7 +6,6 @@ set -e
 cd "$HOME/github/gcsfuse/perfmetrics/scripts"
 
 echo "Setting up the machine with Docker and Nvidia Driver"
-chmod +x ml_tests/setup_host.sh
 source ml_tests/setup_host.sh
 
 cd "$HOME/github/gcsfuse"
@@ -20,7 +19,6 @@ sudo docker run --gpus all --name=pytorch_automation_container --privileged -d -
 --shm-size=128g pytorch-gcsfuse:latest
 
 # Setup the log_rotation.
-chmod +x perfmetrics/scripts/ml_tests/setup_log_rotation.sh
 source perfmetrics/scripts/ml_tests/setup_log_rotation.sh $HOME/github/gcsfuse/container_artifacts/gcsfuse.log
 
 # Wait for the script completion as well as logs output.