--- a/perfmetrics/scripts/ml_tests/run_image_recognition_models.py
+++ b/perfmetrics/scripts/ml_tests/run_image_recognition_models.py
@@ -136,7 +136,6 @@ def _run_model(directory_name, data_path, data_read_method, ml_model_path, req_f
   end_time = int(time.time())
 
   os.system(f'''cd ..
-            chmod +x populate_metrics.sh
             ./populate_metrics.sh {start_time} {end_time}
             ''')