### Description
Remove chmod calls for running scripts

As in the previous commit, we now have
exec-mode set on all script files, and thus
chmod for runningi them is not needed anymore.

This is on top of [Add exec-mode to script files](https://github.com/GoogleCloudPlatform/gcsfuse/pull/1486) .

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - Ran as presubmit