@gargnitingoogle 2023-11-10
> This PR will need a rebase probably once its dependency #1486 is merged into master.

This is done.