@gargnitingoogle 2023-11-10
@ashmeenkaur @raj-prince the fio perf presubmit tests passed on this PR.

[Build log](https://fusion2.corp.google.com/ci/kokoro/prod:gcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/8beb3bb9-9f93-4919-8b08-d33f2ff38879/log)

+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   |  459.0MiB/s  | 1.19MiB/s  |  34.56MiB/s  |  1.13MiB/s   |
|   PR   |  0.25MiB   | 485.99MiB/s  | 1.05MiB/s  |  35.17MiB/s  |  0.96MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4401.07MiB/s | 76.13MiB/s | 1157.14MiB/s |  79.57MiB/s  |
|   PR   | 48.828MiB  | 4370.97MiB/s | 75.01MiB/s | 1136.26MiB/s |  72.41MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4501.96MiB/s | 33.29MiB/s |  514.7MiB/s  |  32.11MiB/s  |
|   PR   | 976.562MiB | 4054.88MiB/s | 28.8MiB/s  | 681.05MiB/s  |  23.41MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+